+++ entity.h
@@ -7,7 +7,7 @@
 typedef struct entity Entity;
 
 SDL_Rect entity_bbox(struct entity *entity);
-SDL_Rect entity_bbox_after_move(struct entity *entity, int dx, int dy);
+SDL_Rect entity_bbox_after_move(struct entity *entity, float dx, float dy);
 SDL_Rect entity_pos(struct entity *entity);
 
 #endif
